pyproject.toml
[tool.poetry]
name = "rest-project"
version = "0.1.0"
description = ""
authors = ["Konstantin Korovin <ryan337@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.10"
django = "^5.2.6"
pillow = "^11.3.0"
djangorestframework = "^3.16.1"
python-dotenv = "^1.1.1"
django-filter = "^25.1"
shell = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'